queue key is size_t

+++ incs/queue_42.h
@@ -6,7 +6,7 @@
 /*   By: rpinoit <rpinoit@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/09 13:58:18 by rpinoit           #+#    #+#             */
-/*   Updated: 2019/02/12 12:48:17 by rpinoit          ###   ########.fr       */
+/*   Updated: 2019/04/01 23:58:58 by rpinoit          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 
 typedef struct s_qnode
 { 
-    int key; 
+    size_t key; 
     struct s_qnode *next;
 } t_qnode; 
   
@@ -29,11 +29,11 @@ typedef struct s_queue
 
 t_queue *create_queue(void);
 
-t_qnode *new_qnode(int k);
+t_qnode *new_qnode(size_t k);
 
-int de_queue(t_queue *queue);
+size_t de_queue(t_queue *queue);
 
-void en_queue(t_queue *queue, int k);
+void en_queue(t_queue *queue, size_t k);
 
 bool is_queue(t_queue *queue);
